+++ app/views/IduMap/styles.css
@@ -1,0 +1,4 @@
+.page {
+    display: flex;
+    flex-direction: column;
+}
